+++ listfiles/validate.go
